+++ docs/js/metro.js
@@ -42,6 +42,16 @@ String.prototype.isColor = function () {
 return /(^#[0-9A-F]{6}$)|(^#[0-9A-F]{3}$)/i.test(this);
 };
 
+window.secondsToFormattedString = function(time){
+    var hours, minutes, seconds;
+
+    hours = parseInt( time / 3600 ) % 24;
+    minutes = parseInt( time / 60 ) % 60;
+    seconds = time % 60;
+
+    return (hours ? (hours) + ":" : "") + (minutes < 10 ? "0"+minutes : minutes) + ":" + (seconds < 10 ? "0"+seconds : seconds);
+};
+
 Array.prototype.shuffle = function () {
     var currentIndex = this.length, temporaryValue, randomIndex;
 
@@ -2187,6 +2197,348 @@ $.widget("metro.accordion", {
             this._super('_setOption', key, value);
         }
     });
+
+// Source: js/widgets/audio-player.js
+$.widget( "metro.audio" , {
+
+    version: "3.0.14",
+
+    options: {
+        src: false,
+        volume: .5,
+        muted: false,
+        loop: false,
+        preload: false,
+        autoplay: false,
+
+        loopButton: "<span class='mif-loop'></span>",
+        stopButton: "<span class='mif-stop'></span>",
+        playButton: "<span class='mif-play'></span>",
+        pauseButton: "<span class='mif-pause'></span>",
+        muteButton: "<span class='mif-volume-mute'></span>",
+        shuffleButton: "<span class='mif-shuffle'></span>",
+
+        volumeLowButton: "<span class='mif-volume-low'></span>",
+        volumeMediumButton: "<span class='mif-volume-medium'></span>",
+        volumeHighButton: "<span class='mif-volume-high'></span>"
+
+    },
+
+    _create: function () {
+        var that = this, element = this.element, o = this.options;
+
+        this._setOptionsFromDOM();
+
+        this._createPlayer();
+        this._addControls();
+        this._addEvents();
+
+        element.data('audio', this);
+    },
+
+    _createPlayer: function(){
+        var that = this, element = this.element, o = this.options;
+        var audio = element.find("audio");
+
+        element.addClass("audio-player");
+
+        if (audio.length == 0) {
+            audio = $("<audio>").appendTo(element);
+        }
+
+        $.each(['autoplay', 'controls', 'muted', 'loop', 'preload'], function(){
+            audio.removeAttr(this);
+        });
+
+        if (o.src) {
+            audio.attr(src, o.src);
+        }
+
+        if (o.loop) {
+            audio.attr("loop", "loop");
+        }
+
+        if (o.preload) {
+            audio.attr("preload", "auto");
+        }
+
+        if (o.autoplay) {
+            audio.attr("autoplay", "autoplay");
+        }
+
+        audio[0].volume = o.volume;
+        audio[0].muted = o.muted;
+
+        element.data('muted', false);
+        element.data('duration', 0);
+        element.data('played', false);
+    },
+
+    _addControls: function(){
+        var that = this, element = this.element, o = this.options;
+        var controls, play_button, loop_button, stop_button, volume_button, volume_slider, stream_slider, info_box, stream_wrapper, volume_wrapper, shufle_button;
+        var audio = element.find('audio'), audio_obj = audio[0];
+
+        controls = $("<div>").addClass("controls").appendTo(element);
+
+        stream_wrapper = $("<div/>").addClass('stream-wrapper').appendTo(controls);
+        stream_slider = $("<div/>").addClass('slider stream-slider').appendTo(stream_wrapper);
+        stream_slider.slider({
+            showHint: true,
+            animate: false,
+            markerColor: 'bg-red',
+            completeColor: 'bg-cyan',
+            onStartChange: function(){
+                audio_obj.pause();
+            },
+            onChanged: function(value, slider){
+                if (audio_obj.seekable.length > 0)
+                    audio_obj.currentTime = (element.data('duration') * value / 100).toFixed(0);
+
+                if (element.data('played') && audio_obj.currentTime >= 0) {
+                    audio_obj.play();
+                }
+            }
+        });
+        stream_slider.data('slider').value(0);
+
+        loop_button = $("<button/>").addClass("square-button control-element loop").html(o.loopButton).appendTo(controls);
+        loop_button.on("click", function(){
+            loop_button.toggleClass('active');
+            if (loop_button.hasClass('active')) {
+                audio.attr("loop", "loop");
+            } else {
+                audio.removeAttr("loop");
+            }
+        });
+
+        play_button = $("<button/>").addClass("square-button control-element play").html(o.playButton).appendTo(controls);
+        play_button.on("click", function(){
+            that._play();
+        });
+
+        stop_button = $("<button/>").addClass("square-button control-element stop").html(o.stopButton).appendTo(controls);
+        stop_button.attr("disabled", true);
+        stop_button.on("click", function(){
+            that._stop();
+        });
+
+        info_box = $("<div/>").addClass('control-element info-box').appendTo(controls);
+        info_box.html("00:00 / 00:00");
+
+        var volume_container = $("<div/>").addClass("place-right").appendTo(controls);
+
+        volume_button = $("<button/>").addClass("square-button control-element volume").html(o.volumeLowButton).appendTo(volume_container);
+        volume_button.on("click", function(){
+            element.data('muted', !element.data('muted'));
+
+            audio_obj.muted = element.data('muted');
+
+            if (element.data('muted')) {
+                volume_button.html(o.muteButton);
+            } else {
+                that._setupVolumeButton();
+            }
+        });
+
+        this._setupVolumeButton();
+
+        volume_wrapper = $("<div/>").addClass('control-element volume-wrapper').appendTo(volume_container);
+        volume_slider = $("<div/>").addClass('slider volume-slider').appendTo(volume_wrapper);
+        volume_slider.slider({
+            showHint: true,
+            animate: false,
+            markerColor: 'bg-red',
+            completeColor: 'bg-green',
+            onChange: function(value, slider){
+                audio_obj.volume = value/100;
+                that._setupVolumeButton();
+            }
+        });
+        volume_slider.data('slider').value(audio_obj.volume * 100);
+
+    },
+
+    _setupVolumeButton: function(){
+        var that = this, element = this.element, o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var controls = element.find('.controls'), volume_button = controls.find('.volume');
+
+        var current_volume = audio_obj.volume;
+        if (current_volume > 0 && current_volume < 0.3) {
+            volume_button.html(o.volumeLowButton);
+        } else if (current_volume >= 0.3 && current_volume < 0.6) {
+            volume_button.html(o.volumeMediumButton);
+        } else if (current_volume >= 0.6 && current_volume <= 1) {
+            volume_button.html(o.volumeHighButton);
+        } else {
+            volume_button.html(o.muteButton);
+        }
+    },
+
+    _addEvents: function(){
+        var that = this, element = this.element, o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var controls = element.find(".controls");
+        var info_box = element.find(".info-box");
+
+        audio.on('loadedmetadata', function(){
+            element.data('duration', audio_obj.duration.toFixed(0));
+            info_box.html("00:00" + " / " + secondsToFormattedString(element.data('duration')) );
+        });
+
+        audio.on("canplay", function(){
+            //preloader.hide();
+            var buffered = audio_obj.buffered.length ? Math.round(Math.floor(audio_obj.buffered.end(0)) / Math.floor(audio_obj.duration) * 100) : 0;
+            that._setBufferSize(buffered);
+        });
+
+        audio.on('progress', function(){
+            var buffered = audio_obj.buffered.length ? Math.round(Math.floor(audio_obj.buffered.end(0)) / Math.floor(audio_obj.duration) * 100) : 0;
+            that._setBufferSize(buffered);
+        });
+
+        audio.on("timeupdate", function(){
+            that._setInfoData();
+            that._setStreamSliderPosition();
+        });
+
+        audio.on("waiting", function(){
+            //preloader.show();
+        });
+
+        audio.on("loadeddata", function(){
+            //preloader.hide();
+        });
+
+        audio.on('ended', function(){
+            that._stop();
+        });
+    },
+
+    _setInfoData: function(){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var info_box = element.find(".controls .info-box");
+        var currentTime = Math.round(audio_obj.currentTime);
+
+        info_box.html(secondsToFormattedString(currentTime) + " / " + secondsToFormattedString(element.data('duration')));
+    },
+
+    _setStreamSliderPosition: function(){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var slider = element.find(".stream-slider").data("slider");
+        var value = Math.round(audio_obj.currentTime * 100 / element.data('duration'));
+        slider.value(value);
+    },
+
+
+    _setBufferSize: function(value){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var slider = element.find(".stream-slider").data("slider");
+        slider.buffer(Math.round(value));
+    },
+
+    _play: function(){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var play_button = element.find(".controls .play");
+        var stop_button = element.find(".controls .stop");
+
+        if (audio_obj.paused) {
+            play_button.html(o.pauseButton);
+            audio_obj.play();
+            stop_button.removeAttr("disabled");
+            element.data('played', true);
+            element.trigger('play');
+        } else {
+            play_button.html(o.playButton);
+            audio_obj.pause();
+            element.data('played', false);
+            element.trigger('pause');
+        }
+    },
+
+    _stop: function(){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0];
+        var stop_button = element.find(".controls .stop");
+        var play_button = element.find(".controls .play");
+
+        audio_obj.pause();
+        audio_obj.currentTime = 0;
+        play_button.html(o.playButton);
+        stop_button.attr("disabled", "disabled");
+        element.data('played', false);
+        element.find(".stream-slider").data('slider').value(0);
+        element.trigger('stop');
+    },
+
+
+    play: function(file, type){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0], source;
+
+        this._stop();
+
+        audio.find("source").remove();
+        source = $("source").attr("src", file);
+        if (type != undefined) {
+            source.attr("type", type);
+        }
+        source.appendTo(audio);
+
+        this._play();
+    },
+
+    pause: function(){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0], play_button = element.find(".play");
+
+        play_button.html(o.playButton);
+        audio_obj.pause();
+        element.data('played', false);
+        element.trigger('pause');
+    },
+
+    resume: function(){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var audio = element.find("audio"), audio_obj = audio[0], play_button = element.find(".play"), stop_button = element.find(".stop");
+
+        play_button.html(o.pauseButton);
+        audio_obj.play();
+        stop_button.removeAttr("disabled");
+        element.data('played', true);
+        element.trigger('play');
+    },
+
+    stop: function(){
+        this._stop();
+    },
+
+    _setOptionsFromDOM: function(){
+        var that = this, element = this.element, o = this.options;
+
+        $.each(element.data(), function(key, value){
+            if (key in o) {
+                try {
+                    o[key] = $.parseJSON(value);
+                } catch (e) {
+                    o[key] = value;
+                }
+            }
+        });
+    },
+
+    _destroy: function () {
+    },
+
+    _setOption: function ( key, value ) {
+        this._super('_setOption', key, value);
+    }
+});
 
 // Source: js/widgets/button-groups.js
 $.widget( "metro.group" , {
@@ -6853,7 +7205,6 @@ $.widget("metro.slider", {
         }
 
         element.children('.marker').on(event_down, function (e) {
-            e.preventDefault();
             that._startMoveMarker(e);
             if (typeof o.onStartChange === 'function') {
                 o.onStartChange();
@@ -6865,6 +7216,8 @@ $.widget("metro.slider", {
                     result.call();
                 }
             }
+            e.preventDefault();
+            e.stopPropagation();
         });
 
         element.on(event_down, function (e) {
@@ -6882,15 +7235,15 @@ $.widget("metro.slider", {
         var event_move = isTouchDevice() ? 'touchmove' : 'mousemove';
         var event_up = isTouchDevice() ? 'touchend' : 'mouseup mouseleave';
 
-        $(element).on(event_move, function (event) {
+        $(document).on(event_move, function (event) {
             that._movingMarker(event);
             if (!element.hasClass('permanent-hint')) {
                 hint.css('display', 'block');
             }
         });
-        $(element).on(event_up, function () {
-            $(element).off('mousemove');
-            $(element).off('mouseup');
+        $(document).on(event_up, function () {
+            $(document).off(event_move);
+            $(document).off(event_up);
             element.data('value', o.position);
             element.trigger('changed', o.position);
             element.trigger('change', o.position);
@@ -6997,7 +7350,7 @@ $.widget("metro.slider", {
         if (o._slider.vertical) {
             var oldSize = this._percToPix(o.position) + o._slider.marker,
                 oldSize2 = o._slider.length - oldSize;
-            size = this._percToPix(value) + o._slider.marker;
+            size = this._percToPix(value) + o._slider.marker / 2;
             size2 = o._slider.length - size;
             this._animate(marker.css('top', oldSize2),{top: size2});
             this._animate(complete.css('height', oldSize),{height: size});
@@ -7008,7 +7361,7 @@ $.widget("metro.slider", {
             }
             if (o.showHint) {
                 hintValue = this._valueToRealValue(value);
-                hint.html(hintValue).css('top', size2 - hint.height()/2 + (element.hasClass('large') ? 8 : 0));
+                hint.html(hintValue).css('top', size2 - marker.height()/2 - hint.height()/4);
             }
         } else {
             size = this._percToPix(value);
@@ -7020,7 +7373,7 @@ $.widget("metro.slider", {
             }
             if (o.showHint) {
                 hintValue = this._valueToRealValue(value);
-                hint.html(hintValue).css({left: size - hint.width() / 2 + (element.hasClass('large') ? 6 : 0)});
+                hint.html(hintValue).css('left', size - marker.width()/2);
             }
         }
     },
@@ -7045,7 +7398,7 @@ $.widget("metro.slider", {
 
     _animate: function (obj, val) {
         var o = this.options;
-
+        //console.log(obj, val);
         if(o.animate) {
             obj.stop(true).animate(val);
         } else {
@@ -7055,11 +7408,12 @@ $.widget("metro.slider", {
 
     _pixToPerc: function (valuePix) {
         var valuePerc;
-        valuePerc = valuePix * this.options._slider.ppp;
+        valuePerc = (valuePix < 0 ? 0 : valuePix )* this.options._slider.ppp;
         return Math.round(this._correctValue(valuePerc));
     },
 
     _percToPix: function (value) {
+        ///console.log(this.options._slider.ppp, value);
         if (this.options._slider.ppp === 0) {
             return 0;
         }
@@ -7111,10 +7465,11 @@ $.widget("metro.slider", {
     _createSlider: function(){
         var element = this.element,
             o = this.options,
-            complete, marker, hint, buffer;
+            complete, marker, hint, buffer, back;
 
         element.html('');
 
+        back = $("<div/>").addClass("slider-backside").appendTo(element);
         complete = $("<div/>").addClass("complete").appendTo(element);
         buffer = $("<div/>").addClass("buffer").appendTo(element);
         marker = $("<a/>").addClass("marker").appendTo(element);
@@ -7125,9 +7480,9 @@ $.widget("metro.slider", {
 
         if (o.color !== 'default') {
             if (o.color.isColor()) {
-                element.css('background-color', o.color);
+                back.css('background-color', o.color);
             } else {
-                element.addClass(o.color);
+                back.addClass(o.color);
             }
         }
         if (o.completeColor !== 'default') {
@@ -8837,6 +9192,7 @@ $.widget( "metro.video" , {
         controls: true,
         controlsPosition: 'bottom',
         controlsModel: 'full',
+
         loopButton: "<span class='mif-loop'></span>",
         stopButton: "<span class='mif-stop'></span>",
         playButton: "<span class='mif-play'></span>",
@@ -8855,6 +9211,7 @@ $.widget( "metro.video" , {
         loop: false,
         preload: false,
         autoplay: false,
+        muted: false,
         volume:.5,
         logo: false,
 
@@ -8894,7 +9251,7 @@ $.widget( "metro.video" , {
             video = $("<video/>").appendTo(element);
         }
 
-        $.each(['autoplay', 'controls', 'height', 'width', 'loop', 'poster', 'preload'], function(){
+        $.each(['muted', 'autoplay', 'controls', 'height', 'width', 'loop', 'poster', 'preload'], function(){
             video.removeAttr(this);
         });
 
@@ -9029,13 +9386,14 @@ $.widget( "metro.video" , {
     },
 
     _timeToString: function(time){
-        var hours, minutes, seconds;
-
-        hours = parseInt( time / 3600 ) % 24;
-        minutes = parseInt( time / 60 ) % 60;
-        seconds = time % 60;
-
-        return (hours ? (hours) + ":" : "") + (minutes < 10 ? "0"+minutes : minutes) + ":" + (seconds < 10 ? "0"+seconds : seconds);
+        return secondsToFormattedString(time);
+        //var hours, minutes, seconds;
+        //
+        //hours = parseInt( time / 3600 ) % 24;
+        //minutes = parseInt( time / 60 ) % 60;
+        //seconds = time % 60;
+        //
+        //return (hours ? (hours) + ":" : "") + (minutes < 10 ? "0"+minutes : minutes) + ":" + (seconds < 10 ? "0"+seconds : seconds);
     },
 
     _setInfoData: function(){
@@ -9130,8 +9488,9 @@ $.widget( "metro.video" , {
                 if (video_obj.seekable.length > 0)
                     video_obj.currentTime = (element.data('duration') * value / 100).toFixed(0);
 
-                if (element.data('played') && video_obj.currentTime > 0)
+                if (element.data('played') && video_obj.currentTime >= 0) {
                     video_obj.play();
+                }
             }
         });
         stream_slider.data('slider').value(0);