+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * add video player
 * upd slider, add buffer (great for player)
 * fix esc click for full screen video player with mode window
+* add audio player
+* remove large slider
+* redefine slider
 
 ## 3.0.13
 * upd treeview for method addLeaf, now method returned li